+++ extract_msg/ole_writer.py
@@ -1,24 +1,26 @@
+import copy
 import io
 import pathlib
 import re
 
-from typing import List, Optional, Union
+from typing import Dict, Iterator, List, Optional, Tuple, Union
 
 from . import constants
 from .enums import Color, DirectoryEntryType
-from .utils import ceilDiv, inputToMsgPath
+from .utils import ceilDiv, dictGetCasedKey, inputToMsgPath
 from olefile.olefile import OleDirectoryEntry, OleFileIO
 from red_black_dict_mod import RedBlackTree
 
 
-class _DirectoryEntry:
+class DirectoryEntry:
     """
-    Hidden class, will probably be modified later.
+    An internal representation of a stream or storage in the OleWriter.
+    Originals should be inaccessible outside of the class.
     """
     name : str = ''
-    rightChild : '_DirectoryEntry' = None
-    leftChild : '_DirectoryEntry' = None
-    childTreeRoot : '_DirectoryEntry' = None
+    rightChild : 'DirectoryEntry' = None
+    leftChild : 'DirectoryEntry' = None
+    childTreeRoot : 'DirectoryEntry' = None
     stateBits : int = 0
     creationTime : int = 0
     modifiedTime : int = 0
@@ -35,7 +37,7 @@ class _DirectoryEntry:
     startingSectorLocation : int = 0
     color : Color = Color.BLACK
 
-    clsid : bytes = b''
+    clsid : bytes = b'\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00'
     data : bytes = b''
 
     def __init__(self):
@@ -79,13 +81,176 @@ class OleWriter:
     [MS-CFB].
     """
     def __init__(self, rootClsid : bytes = constants.DEFAULT_CLSID):
-        self.__rootClsid = rootClsid
+        self.__rootEntry = DirectoryEntry()
+        self.__rootEntry.name = "Root Entry"
+        self.__rootEntry.type = DirectoryEntryType.ROOT_STORAGE
+        self.__rootEntry.clsid = rootClsid
         # The root entry will always exist, so this must be at least 1.
         self.__dirEntryCount = 1
         self.__dirEntries = {}
         self.__largeEntries = []
         self.__largeEntrySectors = 0
         self.__numMinifatSectors = 0
+
+    def __getContainingStorage(self, path : List[str], entryExists : bool = True, create : bool = False) -> Dict:
+        """
+        Finds the storage dict internally where the entry specified by
+        :param path: would be created. If :param create: is True, missing
+        storages will be created with default settings.
+
+        :param entryExists: If True, throws an error when the requested entry
+            does not yet exist.
+        :param create: If True, creates missing storages with default settings.
+
+        :raises OSError: If :param create: is False and the path could not be
+            found. Also raised if :param entryExists: is True and the requested
+            entry does not exist.
+        :raises ValueError: Tried to access an interal stream or tried to use
+            both the create option and the entryExists option as True.
+
+        :returns: The storage dict that the entry is in.
+        """
+        # Quick check for incompatability between create and entryExists.
+        if create and entryExists:
+            raise ValueError(':param create: and :param entryExists: cannot both be True (an entry cannot exist if it is being created).')
+
+        # Check that the path is not an internal entry. Given the validation on
+        # paths that most functions should do because of the call to
+        # inputToMsgPath, this shouldn't actually be necessary.
+        if any(x.startswith('::') for x in path):
+            raise ValueError('Found internal name in path.')
+
+        _dir = self.__dirEntries
+
+        for index, name in enumerate(path[:-1]):
+            # If no entry in the current stream matches the path, raise an
+            # OSError, *unless* the option to create storages is True.
+            if name.lower() not in map(str.lower, _dir.keys()):
+                if create:
+                    self.addEntry(path[:index + 1], storage = True)
+                else:
+                    raise OSError(f'Entry not found: {name}')
+            _dir = _dir[dictGetCasedKey(_dir, name)]
+
+            # If the current item is not a storage and we have more to the path,
+            # raise an OSError.
+            if not isinstance(_dir, dict):
+                raise OSError('Attempted to access children of a stream.')
+
+        if entryExists and path[-1].lower() not in map(str.lower, _dir.keys()):
+            raise OSError(f'Entry not found: {path[-1]}')
+
+        return _dir
+
+    def __getEntry(self, path : List[str]) -> DirectoryEntry:
+        """
+        Finds and returns an existing DirectoryEntry instance in the writer.
+
+        :raises OSError: If the entry does not exist.
+        :raises ValueError: If access to an internal item is attempted.
+        """
+        _dir = self.__getContainingStorage(path)
+        item = _dir[dictGetCasedKey(_dir, path[-1])]
+        if isinstance(item, dict):
+            return item['::DirectoryEntry']
+        else:
+            return item
+
+    def __modifyEntry(self, entry : DirectoryEntry, **kwargs):
+        """
+        Edits the DirectoryEntry with the data provided. Common code used for
+        :method addEntry: and :method editEntry:.
+
+        :raises ValueError: Some part of the data given to modify the various
+            properties was invalid. See the the listed methods for details.
+        """
+        # Extract the arguments.
+        data = kwargs.get('data')
+        clsid = kwargs.get('clsid')
+        creationTime = kwargs.get('creationTime')
+        modifiedTime = kwargs.get('modifiedTime')
+        stateBits = kwargs.get('stateBits')
+
+        # I don't like that I have repeated if statements for checking each of
+        # the arguments, but I need to make sure nothing changes if something is
+        # invalid.
+        if data is not None:
+            if entry.type is not DirectoryEntryType.STREAM:
+                raise ValueError('Cannot set the data of a storage object.')
+            if not isinstance(data, bytes):
+                raise ValueError('Data must be a bytes instance if set.')
+
+        if clsid is not None:
+            if not isinstance(clsid, bytes):
+                raise ValueError('CLSID must be bytes.')
+            if len(clsid) != 16:
+                raise ValueError('CLSID must be 16 bytes.')
+
+        if creationTime is not None:
+            if entry.type is DirectoryEntryType.STREAM:
+                raise ValueError('Modification of creation time cannot be done on a stream.')
+            if not isinstance(creationTime, int) or creationTime < 0 or creationTime > 0xFFFFFFFFFFFFFFFF:
+                raise ValueError('Creation time must be a positive 8 byte int.')
+
+        if modifiedTime is not None:
+            if entry.type is DirectoryEntryType.STREAM:
+                raise ValueError('Modification of modified time cannot be done on a stream.')
+            if not isinstance(modifiedTime, int) or modifiedTime < 0 or modifiedTime > 0xFFFFFFFFFFFFFFFF:
+                raise ValueError('Modified time must be a positive 8 byte int.')
+
+        if stateBits is not None:
+            if not isinstance(stateBits, int) or stateBits < 0 or stateBits > 0xFFFFFFFF:
+                raise ValueError('State bits must be a positive 4 byte int.')
+
+
+        # Now that all our checks have passed, let's set our data.
+        if data is not None:
+            entry.data = data
+        if clsid is not None:
+            entry.clsid = clsid
+        if creationTime is not None:
+            entry.creationTime = creationTime
+        if modifiedTime is not None:
+            entry.modifiedTime = modifiedTime
+        if stateBits is not None:
+            entry.stateBits = stateBits
+
+    def __recalculateSectors(self) -> None:
+        """
+        Recalculates several of the internal variables used for saving that
+        specify the number of sectors and where things should go.
+        """
+        self.__dirEntryCount = 0
+        self.__numMinifatSectors = 0
+        self.__largeEntries.clear()
+        self.__largeEntrySectors = 0
+
+        count = 0
+        for entry in self.__walkEntries():
+            self.__dirEntryCount += 1
+            if entry.type == DirectoryEntryType.STREAM:
+                if len(entry.data) < 4096:
+                    self.__numMinifatSectors += ceilDiv(len(entry.data), 64)
+                else:
+                    self.__largeEntries.append(entry)
+                    self.__largeEntrySectors += ceilDiv(len(entry.data), 512)
+
+    def __walkEntries(self) -> Iterator[DirectoryEntry]:
+        """
+        Returns a generator that will walk the entires recursively. Each item
+        returned by it will be a DirectoryEntry instance.
+        """
+        toProcess = [self.__dirEntries]
+        yield self.__rootEntry
+
+        while len(toProcess) > 0:
+            for name, item in toProcess.pop(0).items():
+                if not name.startswith('::'):
+                    if isinstance(item, dict):
+                        yield item['::DirectoryEntry']
+                        toProcess.append(item)
+                    else:
+                        yield item
 
     @property
     def __numberOfSectors(self) -> int:
@@ -100,9 +265,27 @@ class OleWriter:
 
     @property
     def __numMinifat(self) -> int:
+        """
+        The number of FAT sectors needed to store the mini FAT.
+        """
         return ceilDiv(self.__numMinifatSectors, 8)
 
-    def _getFatSectors(self):
+    def _cleanupEntries(self) -> None:
+        """
+        Cleans up the node connections by walking the tree and removing
+        references that were added during writing.
+        """
+        self.__largeEntries.clear()
+        for entry in self.__walkEntries():
+            entry.id = -1
+            entry.leftChild = None
+            entry.rightChild = None
+            entry.childTreeRoot = None
+            entry.leftSiblingID = 0xFFFFFFFF
+            entry.rightSiblingID = 0xFFFFFFFF
+            entry.childID = 0xFFFFFFFF
+
+    def _getFatSectors(self) -> Tuple[int, int, int]:
         """
         Returns a tuple containing the number of FAT sectors, the number of
         DIFAT sectors, and the total number of sectors the saved file will have.
@@ -119,19 +302,19 @@ class OleWriter:
 
         return (numFat, numDifat, self.__numberOfSectors + numDifat + numFat)
 
-    def _treeSort(self, startingSector : int) -> List[_DirectoryEntry]:
+    def _treeSort(self, startingSector : int) -> List[DirectoryEntry]:
         """
         Uses red-black trees to sort the internal data in preparation for
         writing the file, returning a list, in order, of the entries to write.
         """
         # First, create the root entry.
-        root = _DirectoryEntry()
-        root.name = "Root Entry"
-        root.type = DirectoryEntryType.ROOT_STORAGE
-        root.clsid = self.__rootClsid
+        root = copy.copy(self.__rootEntry)
+
         # Add the location of the start of the mini stream.
         root.startingSectorLocation = (startingSector + ceilDiv(self.__dirEntryCount, 4) + ceilDiv(self.__numMinifatSectors, 128)) if self.__numMinifat > 0 else 0xFFFFFFFE
         root.streamSize = self.__numMinifatSectors * 64
+        root.childTreeRoot = None
+        root.childID = 0xFFFFFFFF
         entries = [root]
 
         toProcess = [(root, self.__dirEntries)]
@@ -153,9 +336,9 @@ class OleWriter:
                     # the processing list.
                     if isinstance(val, dict):
                         toProcess.append((val['::DirectoryEntry'], val))
-                        entries.append(val['::DirectoryEntry'])
-                    else:
-                        entries.append(val)
+                        val = val['::DirectoryEntry']
+
+                    entries.append(val)
 
                     # Add the data to the tree.
                     tree.add((len(name), name.upper()), val)
@@ -169,23 +352,15 @@ class OleWriter:
             for node in tree.in_order():
                 item = node.value
                 # Set the color immediately.
-                if isinstance(item, dict):
-                    item = item['::DirectoryEntry']
                 item.color = Color.BLACK if node.is_black else Color.RED
+
                 if node.left:
-                    val = node.left.value
-                    if isinstance(val, _DirectoryEntry):
-                        item.leftChild = val
-                    else:
-                        item.leftChild = val['::DirectoryEntry']
+                    item.leftChild = node.left.value
                 else:
                     item.leftChild = None
+
                 if node.right:
-                    val = node.right.value
-                    if isinstance(val, _DirectoryEntry):
-                        item.rightChild = val
-                    else:
-                        item.rightChild = val['::DirectoryEntry']
+                    item.rightChild = node.right.value
                 else:
                     item.rightChild = None
 
@@ -222,6 +397,8 @@ class OleWriter:
 
         :returns: The current sector number after all the data is written.
         """
+        # Recalculate some things needed for saving.
+        self.__recalculateSectors()
         # Since we are going to need these multiple times, get them now.
         numFat, numDifat, totalSectors = self._getFatSectors()
 
@@ -345,7 +522,7 @@ class OleWriter:
         # Finally, return the current sector index for use in other places.
         return numDifat + numFat
 
-    def _writeDirectoryEntries(self, f, startingSector : int) -> List[_DirectoryEntry]:
+    def _writeDirectoryEntries(self, f, startingSector : int) -> List[DirectoryEntry]:
         """
         Writes out all the directory entries. Returns the list generated.
         """
@@ -357,7 +534,7 @@ class OleWriter:
 
         return entries
 
-    def _writeDirectoryEntry(self, f, entry : _DirectoryEntry) -> None:
+    def _writeDirectoryEntry(self, f, entry : DirectoryEntry) -> None:
         """
         Writes the directory entry to the file f.
         """
@@ -373,7 +550,7 @@ class OleWriter:
             if len(x.data) & 511:
                 f.write(b'\x00' * (512 - (len(x.data) & 511)))
 
-    def _writeMini(self, f, entries : List[_DirectoryEntry]) -> None:
+    def _writeMini(self, f, entries : List[DirectoryEntry]) -> None:
         """
         Writes the mini FAT followed by the full mini stream.
         """
@@ -403,40 +580,82 @@ class OleWriter:
         if self.__numMinifatSectors & 7:
             f.write((b'\x00' * 64) * (8 - (self.__numMinifatSectors & 7)))
 
+    def addEntry(self, path, data : bytes = None, storage : bool = False, **kwargs) -> None:
+        """
+        Adds an entry to the OleWriter instance at the path specified, adding
+        storages with default settings where necessary. If the entry is not a
+        storage, :param data: *must* be set.
+
+        :param path: The path to add the entry at. Must not contain a path part
+            that is an already added stream.
+        :param data: The bytes for a stream.
+        :param storage: If True, the entry to add is a storage. Otherwise, the
+            entry is a stream.
+        :param clsid: The CLSID for the stream/storage. Must a a bytes instance
+            that is 16 bytes long.
+        :param creationTime: An 8 byte filetime int. Sets the creation time of
+            the entry. Not applicable to streams.
+        :param modifiedTime: An 8 byte filetime int. Sets the modification time
+            of the entry. Not applicable to streams.
+        :param stateBits: A 4 byte int. Sets the state bits, user-defined flags,
+            of the entry. For a stream, this *SHOULD* be unset.
+
+        :raises OSError: A stream was found on the path before the end.
+        :raises ValueError: Attempts to access an internal item.
+        """
+        path = inputToMsgPath(path)
+        # First, find the current place in our dict to add the item.
+        _dir = self.__getContainingStorage(path, False, True)
+        # Now, check that the item *is not* already in our dict, as that would
+        # cause problems.
+        if path[-1].lower() in map(str.lower, _dir.keys()):
+            raise OSError('Cannot add an entry that already exists.')
+
+        # Create a new entry with basic data and insert it.
+        entry = DirectoryEntry()
+        entry.type = DirectoryEntryType.STORAGE if storage else DirectoryEntryType.STREAM
+        entry.name = path[-1]
+        self.__modifyEntry(entry, data = data, **kwargs)
+        if storage:
+            _dir[path[-1]] = {'::DirectoryEntry': entry}
+        else:
+            _dir[path[-1]] = entry
+
     def addOleEntry(self, path, entry : OleDirectoryEntry, data : Optional[bytes] = None) -> None:
         """
         Uses the entry provided to add the data to the writer.
 
-        :raises ValueError: Tried to add an entry to a path that has not yet
-            been added.
+        :raises OSError: Tried to add an entry to a path that has not yet
+            been added, tried to add as a child of a stream, or tried to add an
+            entry where one already exists under the same name.
         """
-        pathList = inputToMsgPath(path)
+        path = inputToMsgPath(path)
         # First, find the current place in our dict to add the item.
-        _dir = self.__dirEntries
-        while len(pathList) > 1:
-            if pathList[0] not in _dir:
-                # If no entry has been provided already for the directory, that
-                # is considered a fatal error.
-                raise ValueError('Path not found.')
-            _dir = _dir[pathList[0]]
-            pathList.pop(0)
+        
+        _dir = self.__getContainingStorage(path, False)
+        # Now, check that the item *is not* already in our dict, as that would
+        # cause problems.
+        if path[-1].lower() in map(str.lower, _dir.keys()):
+            raise OSError('Cannot add an entry that already exists.')
 
         # Now that we are in the right place, add our data.
-        newEntry = _DirectoryEntry()
+        newEntry = DirectoryEntry()
         if entry.entry_type == DirectoryEntryType.STORAGE:
             # Handle a storage entry.
             # First add the dict to our tree of items.
-            _dir[pathList[0]] = {'::DirectoryEntry': newEntry}
+            _dir[path[-1]] = {'::DirectoryEntry': newEntry}
 
             # Finally, setup the values for the stream.
             newEntry.name = entry.name
             newEntry.type = DirectoryEntryType.STORAGE
             newEntry.clsid = _unClsid(entry.clsid)
             newEntry.stateBits = entry.dwUserFlags
+            newEntry.creationTime = entry.createTime
+            newEntry.modifiedTime = entry.modifyTime
         else:
             # Handle a stream entry.
             # First add the entry to out dict of entries.
-            _dir[pathList[0]] = newEntry
+            _dir[path[-1]] = newEntry
             newEntry.name = entry.name
             newEntry.type = DirectoryEntryType.STREAM
             newEntry.clsid = _unClsid(entry.clsid)
@@ -444,20 +663,64 @@ class OleWriter:
 
             # Finally, handle the data.
             newEntry.data = data or b''
-            if len(newEntry.data) < 4096:
-                self.__numMinifatSectors += ceilDiv(len(newEntry.data), 64)
-            else:
-                self.__largeEntries.append(newEntry)
-                self.__largeEntrySectors += ceilDiv(len(newEntry.data), 512)
 
         self.__dirEntryCount += 1
+
+    def deleteEntry(self, path) -> None:
+        """
+        Deletes the entry specified by :param path:, including all children.
+
+        :raises OSError: If the entry does not exist or a part of the path that
+            is not the last was a stream.
+        :raises ValueError: Attempted to delete an internal data stream.
+        """
+        path = inputToMsgPath(path)
+        # Get the containing storage for the entry.
+        _dir = self.__getContainingStorage(path)
+
+        # The garbage collector will take care of all the loose items, so just
+        # remove the entry. Also, once again we deal with the case insensitive
+        # nature of the path. Even though comparisons are case insensitive, the
+        # path does remember the case used.
+        del _dir[dictGetCasedKey(_dir, path[-1])]
+
+    def editEntry(self, path, **kwargs) -> None:
+        """
+        Used to edit values of an entry by setting the specific kwargs. Set a
+        value to something other than None to set it.
+
+        :param data: The data of a stream. Will error if used for something
+            other than a stream.
+        :param clsid: The CLSID for the stream/storage. Must a a bytes instance
+            that is 16 bytes long.
+        :param creationTime: An 8 byte filetime int. Sets the creation time of
+            the entry. Not applicable to streams.
+        :param modifiedTime: An 8 byte filetime int. Sets the modification time
+            of the entry. Not applicable to streams.
+        :param stateBits: A 4 byte int. Sets the state bits, user-defined flags,
+            of the entry. For a stream, this *SHOULD* be unset.
+
+
+        To convert a 32 character hexadecial CLSID into the bytes for this
+        function, the _unClsid function in the ole_writer submodule can be used.
+
+        :raises OSError: The entry does not exist in the file.
+        :raises TypeError: Attempted to modify the bytes of a storage.
+        :raises ValueError: The type of a parameter was wrong, or the data of a
+            parameter was invalid.
+        """
+        # First, find our entry to edit.
+        entry = self.__getEntry(inputToMsgPath(path))
+
+        # Send it to be modified using the arguments given.
+        self.__modifyEntry(entry, **kwargs)
 
     def fromMsg(self, msg : 'MSGFile') -> None:
         """
         Copies the streams and stream information necessary from the MSG file.
         """
         # Get the root OLE entry's CLSID.
-        self.__rootClsid = _unClsid(msg._getOleEntry('/').clsid)
+        self.__rootEntry.clsid = _unClsid(msg._getOleEntry('/').clsid)
 
         # List both storages and directories, but sort them by shortest length
         # first to prevent errors.
@@ -491,24 +754,179 @@ class OleWriter:
             for x in gen:
                 self.addOleEntry(x, msg._getOleEntry(x, prefix = False), msg._getStream(x, prefix = False))
 
-    def fromOleFile(self, ole : OleFileIO) -> None:
+    def fromOleFile(self, ole : OleFileIO, rootPath = []) -> None:
         """
         Copies all the streams from the proided OLE file into this writer.
-        """
-        # Copy the clsid of the root entry.
-        self.__rootClsid = _unClsid(ole.direntries[0].clsid)
 
-        # Copy all of the other entries.
-        for x in ole.listdir(True, True):
-            entry = ole.direntries[ole._find(x)]
+        NOTE: This method does *not* handle any special rule that may be
+        required by a format that uses the compound binary file format as a base
+        when extracting an embedded directory. For example, MSG files require
+        modification of an embedded properties stream when extracting an
+        embedded MSG file.
+
+        :param rootPath: A path (accepted by olefile.OleFileIO) to the directory
+            to use as the root of the file. If not provided, the file root will
+            be used.
+
+        :raises OSError: If :param rootPath: does not exist in the file.
+        """
+        rootPath = inputToMsgPath(rootPath)
+
+        # Check if the root path is simply the top of the file.
+        if rootPath == []:
+            # Copy the clsid of the root entry.
+            self.__rootEntry.clsid = _unClsid(ole.direntries[0].clsid)
+            paths = {tuple(x): (x, ole.direntries[ole._find(x)]) for x in ole.listdir(True, True)}
+        else:
+            # If it is not the top of the file, we need to do some filtering.
+            # First get the CLSID from the entry the path points to.
+            try:
+                entry = ole.direntries[ole._find(rootPath)]
+                self.__rootEntry.clsid = _unClsid(entry.clsid)
+
+            except OSError as e:
+                if str(e) == 'file not found':
+                    # Get the cause/context for the original exception and use
+                    # it for the new exception. This hides the exception from
+                    # OleFileIO.
+                    context = e.__cause__ or e.__context__
+                    raise OSError('Root path was not found in the OLE file.') from context
+                else:
+                    raise
+
+            paths = {tuple(x[len(rootPath):]): (x, ole.direntries[ole._find(x)])
+                     for x in ole.listdir(True, True) if len(x) > len(rootPath)}
+
+
+        # Copy all of the other entries. Ensure that directories come before
+        # their streams by sorting the paths.
+        for x in sorted(paths.keys()):
+            fullPath, entry = paths[x]
 
             if entry.entry_type == DirectoryEntryType.STREAM:
-                with ole.openstream(x) as f:
+                with ole.openstream(fullPath) as f:
                     data = f.read()
             else:
                 data = None
 
             self.addOleEntry(x, entry, data)
+
+    def getEntry(self, path) -> DirectoryEntry:
+        """
+        Finds and returns a copy of an existing DirectoryEntry instance in the
+        writer. Use this method to check the internal status of an entry.
+
+        :raises OSError: If the entry does not exist.
+        :raises ValueError: If access to an internal item is attempted.
+        """
+        return copy.copy(self.__getEntry(inputToMsgPath(path)))
+
+    def listItems(self, streams = True, storages = False) -> List[List[str]]:
+        """
+        Returns a list of the specified items currently in the writter.
+
+        :param streams: If True, includes the path for each stream in the list.
+        :param storages: If True, includes the path for each storage in the
+            list.
+        """
+        # We are actually abusing the walk function a bit here to life much
+        # easier. The way we do this is to look at the current directory that
+        # the walk function is giving information about and then deciding what
+        # parts of it we want to use. Once we have all the paths created, we
+        # will then sort and return it to give an output similar, if not
+        # identical, to OleFileIO.listdir. The mentioned method sorts keeping
+        # case in mind.
+        if not streams and not storages:
+            return []
+
+        paths = []
+        for currentDir, stor, stre in self.walk():
+            if storages:
+                for name in stor:
+                    paths.append(currentDir + [name])
+            if streams:
+                for name in stre:
+                    paths.append(currentDir + [name])
+
+        paths.sort()
+        return paths
+
+    def renameEntry(self, path, newName : str) -> None:
+        """
+        Changes the name of an entry, leaving it in it's current position.
+
+        :raises OSError: If the entry does not exist or an entry with the new
+            name already exists,
+        :raises ValueError: If access to an internal item is attempted or the
+            new name provided is invalid.
+        """
+        # First, validate the new name.
+        if not isinstance(newName, str):
+            raise ValueError('New name must be a string.')
+        if constants.RE_INVALID_OLE_PATH.search(newName):
+            raise ValueError('Invalid character(s) in new name. Must not contain the following characters: \\//!:')
+        if len(newName) > 31:
+            raise ValueError('New name must be less than 32 characters.')
+
+        # Get the storage for our entry. Entry *must* exist.
+        _dir = self.__getContainingStorage(inputToMsgPath(path))
+
+        # See if an item in the storage already has that new name.
+        if newName.lower() in map(str.lower, _dir.keys()):
+            raise OSError('An entry with the new name already exists.')
+
+        # Get the original name.
+        originalName = dictGetCasedKey(_dir, path[-1])
+
+        # Get the entry to change.
+        entry = _dir[originalName]
+        if isinstance(entry, dict):
+            dirData = entry
+            entry = entry['::DirectoryEntry']
+        else:
+            dirData = None
+
+        # Change the name on the entry first.
+        entry.name = newName
+
+        # Now, we need to remove the item from the current storage and add it
+        # back with the new name.
+        del _dir[originalName]
+
+        if dirData is None:
+            _dir[newName] = entry
+        else:
+            _dir[newName] = dirData
+
+    def walk(self) -> Iterator[Tuple[List[str], List[str], List[str]]]:
+        """
+        Functional equivelent to :function os.walk:, but for going over the file
+        structure of the OLE file to be written. Unlike :function os.walk:, it
+        takes no arguments.
+
+        :returns: A tuple of three lists. The first is the path, as a list of
+            strings, for the directory (or an empty list for the root), the
+            second is a list of the storages in the current directory, and the
+            last is a list of the streams. Streams and storages are sorted
+            caselessly.
+        """
+        toProcess = [([], self.__dirEntries)]
+
+        # Go through the toProcess list, removing the last item every time to
+        # mimic the behavior of os.walk.
+        while toProcess:
+            currentDir, dirDict = toProcess.pop()
+            storages = []
+            streams = []
+            for name in sorted(dirDict.keys(), key = str.lower):
+                if not name.startswith('::'):
+                    if isinstance(dirDict[name], dict):
+                        storages.append(name)
+                        toProcess.append((currentDir + [name], dirDict[name]))
+                    else:
+                        streams.append(name)
+
+            yield (currentDir, storages, streams)
 
     def write(self, path) -> None:
         """
@@ -527,12 +945,15 @@ class OleWriter:
         # Make sure we close the file after everything, especially if there is
         # an error.
         try:
-            ### First we need to write the header.
+            # Write each section, transferring data between functions where
+            # necessary.
             offset = self._writeBeginning(f)
             entries = self._writeDirectoryEntries(f, offset)
             self._writeMini(f, entries)
             self._writeFinal(f)
         finally:
+            self._cleanupEntries()
+
             if opened:
                 f.close()